Refactor all interger Ids to Long and add Delete functionality to TodoController

+++ src/main/java/com/example/todorestful/dao/JdbcTodoListDAO.java
@@ -33,7 +33,7 @@ public class JdbcTodoListDAO implements TodoListDAO {
 		parameters.put("description", entity.getDescription());
 		parameters.put("user_id", entity.getUserId());
 		
-		Integer newId = jdbcInsert.executeAndReturnKey(parameters).intValue();
+		Long newId = jdbcInsert.executeAndReturnKey(parameters).longValue();
 		entity.setId(newId);
 		return entity;
 	}
